Google API example
This is a working example of the Google Place API. Here, we are able to search addresses with: Zip code, City name, Country name, ect. I have limited it to searching via address because Google provided extra search fields, which seemed unnecessary. I have commented them out in the code if you would like to play with it. In addition, the Google Maps aspect is included within this example. When we search for an address, the Map takes the user to the location.